Component public interface test helpers
* Add include directory to cpp tests for modulo_components with
declarations for the public interface helper classes

* Merge ...ParameterPublicInterface into PublicInterface and
remove forward declaration from ComponentInterface

* Remove class declarations from test sources
